updates, refactor ExpressionAttributes
